script/requirements.txt
# certifi==2021.10.8
# charset-normalizer==2.0.10
# click==8.0.3
# colorama==0.4.4
Flask==1.1.4
# Flask-Assets==2.0
# Flask-Script==2.0.6
Flask-Script
Flask-SQLAlchemy==2.5.1
# greenlet==1.1.2
# idna==3.3
# itsdangerous==2.0.1
Jinja2
# MarkupSafe==2.0.1
PyMySQL==1.0.2
PyYAML==6.0
requests==2.27.1
SQLAlchemy==1.4.31
urllib3==1.26.8
# webassets==2.0
Werkzeug < 2.0.0
